refactored structure, deleted old files
